@@ -1,25 +1,39 @@
 {
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
-    "start": "expo start",
+    "start": "expo start -c",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "eject": "expo eject"
   },
   "dependencies": {
-    "expo": "^32.0.6",
-    "react": "16.5.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-32.0.0.tar.gz",
-    "react-native-awesome-alert": "^0.4.2",
-    "react-native-countdown-component": "^2.4.0",
-    "react-native-elements": "^1.1.0",
-    "react-native-leaderboard": "^1.0.5",
-    "react-native-modal": "^10.0.0",
-    "react-native-sound": "^0.10.12",
-    "react-navigation": "^3.8.1"
+    "@expo/vector-icons": "^10.2.0",
+    "@react-native-community/masked-view": "0.1.6",
+    "expo": "^37.0.12",
+    "react": "^16.13.1",
+    "react-native": "^0.62.2",
+    "react-native-awesome-alert": "^0.4.5",
+    "react-native-countdown-component": "^2.7.1",
+    "react-native-elements": "^2.0.0",
+    "react-native-gesture-handler": "^1.6.1",
+    "react-native-leaderboard": "^1.0.6",
+    "react-native-modal": "^11.5.6",
+    "react-native-reanimated": "~1.7.0",
+    "react-native-safe-area-context": "0.7.3",
+    "react-native-safe-area-view": "^1.1.1",
+    "react-native-screens": "~2.2.0",
+    "react-native-sound": "^0.11.0",
+    "react-native-vector-icons": "^6.6.0",
+    "react-navigation": "^4.3.9",
+    "react-navigation-stack": "^2.5.1",
+    "socket.io-client": "^2.3.0"
   },
   "devDependencies": {
-    "babel-preset-expo": "^5.0.0"
+    "babel-preset-expo": "^8.2.1",
+    "metro-react-native-babel-preset": "^0.58.0"
+  },
+  "resolutions": {
+    "metro-config": "0.58.0"
   },
   "private": true
 }
