@@ -1,7 +1,7 @@
 import React from 'react';
-import { StyleSheet, View, Button } from 'react-native';
-import { Ionicons, FontAwesome, Entypo } from '@expo/vector-icons'; // 6.2.2
-import {createStackNavigator, createAppContainer, NavigationActions} from 'react-navigation';
+// import { Ionicons, FontAwesome, Entypo } from '@expo/vector-icons'; // 6.2.2
+import {createAppContainer} from 'react-navigation';
+import {createStackNavigator} from 'react-navigation-stack';
 import Single from './app/screen/Single';
 import Home from './app/screen/Home';
 import HighScore from './app/screen/HighScore';
@@ -24,7 +24,7 @@ const MainNavigator = createStackNavigator(
     initialRouteName: 'Home',
     defaultNavigationOptions: {
       //headerMode: 'none',
-      header:null,
+      headerShown: false,
       navigationOptions: {
         headerVisible: false,
       }
