@@ -12,7 +12,7 @@ class Home extends React.Component{
         BackHandler.addEventListener('hardwareBackPress', this.handleBackButton);
     }
     
-    componentWillUnmount() {
+    UNSAFE_componentWillUnmount() {
         BackHandler.removeEventListener('hardwareBackPress', this.handleBackButton);
     }
 
